Reorganise layout components
Move main from routes to modules and renamed to MainLayout
Move minimal from routes to modules renamed to MinimalLayout
