+++ web/reactjs/src/modules/MainLayout/components/Footer/index.js
