+++ web/reactjs/src/modules/MainLayout/components/Sidebar/components/Profile/Profile.js
