+++ web/reactjs/src/modules/MinimalLayout/components/Topbar/index.js
