+++ web/reactjs/src/modules/MinimalLayout/components/index.js
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { default as Topbar } from './Topbar';
+export { default } from './MinimalLayout';
