+++ web/reactjs/src/modules/index.js
@@ -15,3 +15,5 @@
 export { default as Auth } from './Auth';
 export { default as Dashboard } from './Dashboard';
 export { default as NotFound } from './NotFound';
+export { default as MainLayout } from './MainLayout';
+export { default as MinimalLayout } from './MinimalLayout';
